feat: remove runAsUser and add runAsNonRoot
Remove the runAsUser setting so that we run as the default image
USER (coder, uid 1000) in unrestricted Kubernetes environments.
In restricted security environments, such as OpenShift with the
restricted Security Context Constraint, we will run as the uid
for the namespace.

This also applies additional security constraints, preventing
privilege escalation for the migration container and running
as a non-root user.